+++ relay-server/src/processing/attachments/mod.rs
@@ -7,7 +7,6 @@ use crate::envelope::{EnvelopeHeaders, Item, ItemType, Items};
 use crate::managed::{Counted, Managed, ManagedEnvelope, OutcomeError, Quantities, Rejected};
 use crate::processing::utils::attachments;
 use crate::processing::{self, CountRateLimited, Output, QuotaRateLimiter};
-#[cfg(feature = "processing")]
 use crate::services::outcome::DiscardReason;
 use crate::services::outcome::Outcome;
 use crate::statsd::RelayCounters;
@@ -23,7 +22,6 @@ pub enum Error {
     RateLimited(RateLimits),
 
     /// The envelope did not contain an event ID.
-    #[cfg(feature = "processing")]
     #[error("missing event ID")]
     NoEventId,
 }
@@ -37,8 +35,7 @@ impl OutcomeError for Error {
                 let reason_code = limits.longest().and_then(|limit| limit.reason_code.clone());
                 Some(Outcome::RateLimited(reason_code))
             }
-            #[cfg(feature = "processing")]
-            Self::NoEventId => Some(Outcome::Invalid(DiscardReason::Internal)),
+            Self::NoEventId => Some(Outcome::Invalid(DiscardReason::InvalidEventId)),
         };
         (outcome, self)
     }
@@ -98,11 +95,18 @@ impl processing::Processor for AttachmentProcessor {
         mut attachments: Managed<Self::Input>,
         ctx: processing::Context<'_>,
     ) -> Result<processing::Output<Self::Output>, Rejected<Self::Error>> {
+        let has_event_id = attachments.headers.event_id().is_some();
+
         // Temporary counter to figure out which SDKs are still sending standalone attachments.
         relay_statsd::metric!(
             counter(RelayCounters::StandaloneAttachment) += 1,
-            sdk = client_name_tag(attachments.headers.meta().client_name())
+            sdk = client_name_tag(attachments.headers.meta().client_name()),
+            has_event_id = has_event_id.to_string()
         );
+
+        if !has_event_id {
+            return Err(attachments.reject_err(Error::NoEventId));
+        }
 
         attachments::validate_attachments(&mut attachments, |a| &mut a.attachments, ctx);