+++ relay-server/src/statsd.rs
@@ -1024,7 +1024,20 @@ pub enum RelayCounters {
     ///
     /// This metric is tagged with:
     /// - `sdk`: low-cardinality client name
+    /// - `has_event_id`: whether the envelope contained an event ID
     StandaloneAttachment,
+    /// Amount of user reports processed.
+    ///
+    /// This metric is tagged with:
+    /// - `sdk`: low-cardinality client name
+    /// - `has_event_id`: whether the envelope contained an event ID
+    UserReport,
+    /// Amount of replays processed.
+    ///
+    /// This metric is tagged with:
+    /// - `sdk`: low-cardinality client name
+    /// - `has_event_id`: whether the envelope contained an event ID
+    Replay,
 }
 
 impl CounterMetric for RelayCounters {
@@ -1091,6 +1104,8 @@ impl CounterMetric for RelayCounters {
             RelayCounters::ErrorProcessed => "event.error.processed",
             RelayCounters::TraceMetricNilTraceId => "trace_metric.nil_trace_id",
             RelayCounters::StandaloneAttachment => "processing.standalone_attachment",
+            RelayCounters::UserReport => "processing.user_report",
+            RelayCounters::Replay => "processing.replay",
         }
     }
 }